Second Commit
Utility for the creation of test statistics records, that represents
usage events in the context of the software called "DSpace",
(https://duraspace.org/dspace/) used for the creation of institutional
repositories.

This utility creates fake records that simulates the use over the
objects (COMMUNITY, COLLECTION, ITEM, BITSTREAM) in the repository. For
now, the only events created are the "visits" to the pages of this
objects and the "downloads"of the bitstreams.